Refactor data representation
