+++ README.md
@@ -6,11 +6,11 @@ The idea for this scrapper comes from the following [repo](https://github.com/pl
 
 By running the scrapper, you will scrap every data ([sorted per date](https://nuforc.org/webreports/ndxevent.html)) from the [NUFORC website](https://nuforc.org/).
 
-Data scrapped is stored inside `data.csv` file inside the project directory. If you don't want to run the scrapper, you can download the `dataset.csv` file which may be (and surely is) out of date. 
+ReportRepresentation.Data scrapped is stored inside `data.csv` file inside the project directory. If you don't want to run the scrapper, you can download the `dataset.csv` file which may be (and surely is) out of date. 
 
-### Data structure
+### ReportRepresentation.Data structure
 
-Data inside the CSV file has the same structure as any report webpage (for instance, this [one](https://nuforc.org/webreports/ndxe202205.html)). That is to say :
+ReportRepresentation.Data inside the CSV file has the same structure as any report webpage (for instance, this [one](https://nuforc.org/webreports/ndxe202205.html)). That is to say :
 
 | Date / Time | City | State | Country | Shape | Duration | Summary | Posted | Images |
 |-------------| ---- | ----- | ------- | ----- | -------- | ------- | ------ |--------|
